The C programming Started

+++ basic/temp.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 main() {
-  int fahr, celsius;
+  float fahr, celsius;
   int lower, upper, step;
 
   lower = 0;
@@ -11,7 +11,7 @@ main() {
   fahr = lower;
   while (fahr <= upper) {
     celsius = 5 * (fahr-32) / 9;
-    printf("%d\t%d\n",fahr, celsius);
+    printf("%f\t%f\n",fahr, celsius);
     fahr = fahr + step;
   }
 }
